requirements.txt
asgiref==3.4.1
autocorrect==2.6.0
autopep8==1.6.0
certifi==2021.10.8
click==8.0.3
cloudinary==1.28.0
colorama==0.4.4
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
joblib==1.1.0
mysqlclient==2.1.0
nltk==3.6.5
numpy==1.21.4
Pillow==8.4.0
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
regex==2021.11.10
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tqdm==4.62.3
urllib3==1.26.7
